requirements.txt
torch==2.1.2
transformers==4.39.3
accelerate==0.28.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
runpod>=1.5.0
psutil>=5.9.0
safetensors>=0.3.0
huggingface-hub>=0.16.0